Azure IoT Hub UWP SDK
Microsoft Azure IoTHub UWP SDK facilitates the service endpoint to send
and receive cloud messages to the device end points.
The SDK can be used for the universal applications developed in c# for
windows 10 OS.
This library uses AMQPNetlite to communicate with the Azure IoT hub. A
sample application is provided to know the usage of the APIs.